nicht erneut, wenn lösung bereits existiert

+++ weniger-krumme-touren/mip/java_interface.py
@@ -5,6 +5,7 @@ java_path = "weniger-krumme-touren/build/SimulatedAnnealing.jar"
 
 
 def solveTA(path):
-    os.system(f"java -cp weniger-krumme-touren/build touren.SimulatedAnnealing {path}")
+    if not os.path.exists(path + ".solution"):
+        os.system(f"java -cp weniger-krumme-touren/build touren.SimulatedAnnealing {path}")
     with open(path + ".solution") as f:
         return tuple(map(int, f.readline()[1:-1].split(", ")))
